requirements.txt
djangorestframework==3.12.4
django-crispy-forms